a bit of cleanup and interface manipulation
