--- examples/services/rabbit2/build.go
@@ -1,21 +1,0 @@
-package rabbit
-
-import (
-	"github.com/simiancreative/simiango/meta"
-	"github.com/simiancreative/simiango/service"
-)
-
-func Build(
-	requestID meta.RequestId,
-	rawHeaders service.RawHeaders,
-	rawBody service.RawBody,
-	rawParams service.RawParams,
-) (
-	service.TPL, error,
-) {
-	// create service instance
-	service := rabbitService{
-		Hello: "Sweetie",
-	}
-	return &service, nil
-}
