--- examples/services/rabbit2/config.go
@@ -1,17 +1,0 @@
-package rabbit
-
-import (
-	"github.com/simiancreative/simiango/messaging/amqp"
-	"github.com/simiancreative/simiango/service"
-)
-
-var Config = service.Config{
-	Key:   "hello",
-	Build: Build,
-}
-
-// dont forget to import your package in your main.go for initialization
-// _ "path/to/project/stream"
-func init() {
-	amqp.AddService(Config)
-}
