--- examples/services/rabbit2/result.go
@@ -1,11 +1,0 @@
-package rabbit
-
-import ()
-
-type rabbitService struct {
-	Hello string `json:"hello"`
-}
-
-func (s *rabbitService) Result() (interface{}, error) {
-	return nil, nil
-}
