+++ examples/services/sample/config.go
@@ -1,6 +1,7 @@
 package sample
 
 import (
+	"github.com/simiancreative/simiango/server"
 	"github.com/simiancreative/simiango/service"
 )
 
@@ -9,4 +10,10 @@ var Config = service.Config{
 	Method:    "POST",
 	Path:      "/sample/:id",
 	Build:     Build,
+}
+
+// dont forget to import your package in your main.go for initialization
+// _ "path/to/project/stream"
+func init() {
+	server.AddService(Config)
 }
